+++ CHANGES.md
@@ -13,7 +13,9 @@
 * Fixed crash when removing USD prims in certain order.
 * Fixed issue where Cesium ion session would not resume on reload.
 * The movie capture tool now waits for tilesets to complete loading before it captures a frame.
-* Added support for WMS raster overlays.
+* Added support for Web Map Service (WMS) raster overlays.
+* Added support for Tile Map Service (TMS) raster overlays.
+* Added support for Web Map Tile Service (WMTS) raster overlays.
 * Added raster overlay options: `maximumScreenSpaceError`, `maximumTextureSize`, `maximumSimultaneousTileLoads`, `subTileCacheBytes`.
 * Changed to use UrlAssetAccessor from vsgCs, which does a better job of managing libcurl handles than the existing HttpAssetAccessor.